Fixed one unit test
