+++ src/main/authentication/authenticationHandler.ts
@@ -1,10 +1,10 @@
 const authenticationConfig = require('../authentication/authentication-config.json');
-import {verifiedRoles} from '../authentication/mediaAuthentication';
 
 export const adminAccountCreationRoles = ['SYSTEM_ADMIN', 'INTERNAL_SUPER_ADMIN_CTSC', 'INTERNAL_SUPER_ADMIN_LOCAL'];
 export const manualUploadRoles = ['SYSTEM_ADMIN', 'INTERNAL_SUPER_ADMIN_CTSC', 'INTERNAL_SUPER_ADMIN_LOCAL', 'INTERNAL_ADMIN_CTSC', 'INTERNAL_ADMIN_LOCAL'];
 export const mediaAccountCreationRoles = ['INTERNAL_SUPER_ADMIN_CTSC', 'INTERNAL_ADMIN_CTSC'];
 export const allAdminRoles = ['SYSTEM_ADMIN', 'INTERNAL_SUPER_ADMIN_CTSC', 'INTERNAL_SUPER_ADMIN_LOCAL', 'INTERNAL_ADMIN_CTSC', 'INTERNAL_ADMIN_LOCAL'];
+export const verifiedRoles = ['VERIFIED'];
 
 export function checkRoles(req, roles): boolean {
   if(req.user) {
@@ -19,23 +19,27 @@ export function checkRoles(req, roles): boolean {
   return false;
 }
 
+export function isPermittedMedia(req: any, res, next) {
+  return checkAuthenticatedMedia(req, res, next, verifiedRoles);
+}
+
 export function isPermittedAdmin(req: any, res, next) {
-  return checkAuthenticated(req, res, next, allAdminRoles);
+  return checkAuthenticatedAdmin(req, res, next, allAdminRoles);
 }
 
 export function isPermittedAccountCreation(req: any, res, next){
-  return checkAuthenticated(req, res, next, adminAccountCreationRoles);
+  return checkAuthenticatedAdmin(req, res, next, adminAccountCreationRoles);
 }
 
 export function isPermittedManualUpload(req: any, res, next) {
-  return checkAuthenticated(req, res, next, manualUploadRoles);
+  return checkAuthenticatedAdmin(req, res, next, manualUploadRoles);
 }
 
 export function isPermittedMediaAccount(req: any, res ,next) {
-  return checkAuthenticated(req, res, next, mediaAccountCreationRoles);
+  return checkAuthenticatedAdmin(req, res, next, mediaAccountCreationRoles);
 }
 
-export function checkAuthenticated(req: any, res, next, roles: string[]): boolean {
+export function checkAuthenticatedAdmin(req: any, res, next, roles: string[]): boolean {
   if (checkRoles(req, roles)) {
     req.user.isAdmin = true;
     return next();
@@ -48,3 +52,15 @@ export function checkAuthenticated(req: any, res, next, roles: string[]): boolea
     res.redirect('/login?p=' + authenticationConfig.ADMIN_POLICY);
   }
 }
+
+export function checkAuthenticatedMedia(req: any, res, next, roles: string[]): boolean {
+  if (checkRoles(req, roles)) {
+    return next();
+  } else if (checkRoles(req, allAdminRoles)) {
+    res.redirect('/admin-dashboard');
+  } else {
+    res.redirect('/login?p=' + authenticationConfig.POLICY);
+  }
+}
+
+
